+++ src/DoublyLinkedList.sol
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.13;
+
 import "suave-std/suavelib/Suave.sol";
 import "forge-std/console.sol";
 
@@ -18,7 +19,7 @@ contract DoublyLinkedList {
     }
 
     struct Node {
-        uint val;
+        uint256 val;
         Suave.DataId parent;
         Suave.DataId child;
     }
@@ -35,18 +36,14 @@ contract DoublyLinkedList {
     function initLinkedList() external returns (bytes memory) {
         Suave.DataId _headId = buildHeadTailPointer();
         Suave.DataId _tailId = buildHeadTailPointer();
-        return
-            abi.encodeWithSelector(this.setLinkedList.selector, _headId, _tailId);
+        return abi.encodeWithSelector(this.setLinkedList.selector, _headId, _tailId);
     }
 
     // CALLBACKS
 
     function nullCallback() public payable {}
 
-    function setLinkedList(
-        Suave.DataId _headId,
-        Suave.DataId _tailId
-    ) public payable {
+    function setLinkedList(Suave.DataId _headId, Suave.DataId _tailId) public payable {
         require(Suave.DataId.unwrap(headId) == Suave.DataId.unwrap(nullId), "Already initialized");
         headId = _headId;
         tailId = _tailId;
@@ -55,12 +52,7 @@ contract DoublyLinkedList {
     // CONFIDENTIAL STORE HELPER FUNCTIONS
 
     function buildHeadTailPointer() internal returns (Suave.DataId) {
-        Suave.DataRecord memory record = Suave.newDataRecord(
-            0,
-            addressList,
-            addressList,
-            "headTailPointer"
-        );
+        Suave.DataRecord memory record = Suave.newDataRecord(0, addressList, addressList, "headTailPointer");
         writeRef(record.id, nullId);
         return record.id;
     }
@@ -69,21 +61,13 @@ contract DoublyLinkedList {
      * @notice Returns the DataId for the head or tail of the linked list.
      * @param headTailId Must be one of public state headId or tailId.
      */
-    function getRef(
-        Suave.DataId headTailId
-    ) internal view returns (Suave.DataId) {
-        bool eqHead = Suave.DataId.unwrap(headTailId)==Suave.DataId.unwrap(headId);
-        bool eqTail = Suave.DataId.unwrap(headTailId)==Suave.DataId.unwrap(tailId);
+    function getRef(Suave.DataId headTailId) internal view returns (Suave.DataId) {
+        bool eqHead = Suave.DataId.unwrap(headTailId) == Suave.DataId.unwrap(headId);
+        bool eqTail = Suave.DataId.unwrap(headTailId) == Suave.DataId.unwrap(tailId);
         require(eqHead || eqTail, "Invalid headTailId");
 
-        bytes memory value = Suave.confidentialRetrieve(
-            headTailId,
-            "suavedll:v0:headTailPointer"
-        );
-        HeadTailPointer memory headTailPointer = abi.decode(
-            value,
-            (HeadTailPointer)
-        );
+        bytes memory value = Suave.confidentialRetrieve(headTailId, "suavedll:v0:headTailPointer");
+        HeadTailPointer memory headTailPointer = abi.decode(value, (HeadTailPointer));
         return headTailPointer.ref;
     }
 
@@ -98,14 +82,8 @@ contract DoublyLinkedList {
      * @param nodeId DataId for the node to be retrieved.
      */
     function getNode(Suave.DataId nodeId) internal returns (Node memory node) {
-        require(
-            Suave.DataId.unwrap(nodeId) != Suave.DataId.unwrap(nullId),
-            "Node does not exist"
-        );
-        bytes memory value = Suave.confidentialRetrieve(
-            nodeId,
-            "suavedll:v0:node"
-        );
+        require(Suave.DataId.unwrap(nodeId) != Suave.DataId.unwrap(nullId), "Node does not exist");
+        bytes memory value = Suave.confidentialRetrieve(nodeId, "suavedll:v0:node");
         node = abi.decode(value, (Node));
         return node;
     }
@@ -124,17 +102,8 @@ contract DoublyLinkedList {
      * @notice Takes in raw fields for a node and creates a new node
      * @dev Used only for node creation, to update existing nodes 'writeNode' should be used
      */
-    function createNode(
-        uint val,
-        Suave.DataId parent,
-        Suave.DataId child
-    ) internal returns (Suave.DataId) {
-        Suave.DataRecord memory record = Suave.newDataRecord(
-            0,
-            addressList,
-            addressList,
-            "node"
-        );
+    function createNode(uint256 val, Suave.DataId parent, Suave.DataId child) internal returns (Suave.DataId) {
+        Suave.DataRecord memory record = Suave.newDataRecord(0, addressList, addressList, "node");
         Node memory node = Node(val, parent, child);
         bytes memory value = abi.encode(node);
         Suave.confidentialStore(record.id, "suavedll:v0:node", value);
@@ -146,7 +115,7 @@ contract DoublyLinkedList {
     /**
      * @notice Push a uint to the head of the linked list
      */
-    function pushHead(uint val) external returns (bytes memory) {
+    function pushHead(uint256 val) external returns (bytes memory) {
         // Steps:
         // Create a new node for this value
         // Set that node's child to be the previous head
@@ -172,7 +141,7 @@ contract DoublyLinkedList {
     /**
      * @notice Push a uint to the tail of the linked list
      */
-    function pushTail(uint val) external returns (bytes memory) {
+    function pushTail(uint256 val) external returns (bytes memory) {
         // Steps:
         // Create a new node for this value
         // Set that node's parent to be the current tail
@@ -207,9 +176,7 @@ contract DoublyLinkedList {
         // If we have no head this will revert
         Node memory headNode = getNode(_headId);
 
-        if (
-            Suave.DataId.unwrap(headNode.child) == Suave.DataId.unwrap(nullId)
-        ) {
+        if (Suave.DataId.unwrap(headNode.child) == Suave.DataId.unwrap(nullId)) {
             writeRef(headId, nullId);
             writeRef(tailId, nullId);
         } else {
@@ -235,9 +202,7 @@ contract DoublyLinkedList {
         // If we have no tail, this will revert
         Node memory tailNode = getNode(_tailId);
 
-        if (
-            Suave.DataId.unwrap(tailNode.parent) == Suave.DataId.unwrap(nullId)
-        ) {
+        if (Suave.DataId.unwrap(tailNode.parent) == Suave.DataId.unwrap(nullId)) {
             writeRef(headId, nullId);
             writeRef(tailId, nullId);
         } else {
